update the docs to split stuff into atomic elements
