--- docs/examples/JustifiedContainer.jsx
@@ -1,9 +1,0 @@
-var example = (
-  <UIToolkit.JustifiedContainer>
-    <div><UIToolkit.Button size="block">foo</UIToolkit.Button></div>
-    <div><UIToolkit.Button size="block">bar</UIToolkit.Button></div>
-    <div><UIToolkit.Button size="block">baz</UIToolkit.Button></div>
-  </UIToolkit.JustifiedContainer>
-);
-
-React.render(example, mountNode);
