+++ docs/src/Nav.jsx
@@ -15,23 +15,28 @@ var Nav = React.createClass({
           </li>
 
           <li>
-            <a href="#components">Components</a>
+            <a href="#atoms">Atoms</a>
             <ul>
-              <li><a href="#alert">Alert</a></li>
               <li><a href="#anchor">Anchor</a></li>
               <li><a href="#button">Buttons</a></li>
-              <li><a href="#countdown">Countdown</a></li>
               <li><a href="#flags">Flags</a></li>
               <li><a href="#image">Image</a></li>
               <li><a href="#input">Input</a></li>
-              <li><a href="#justified-container">Justified Container</a></li>
               <li><a href="#lists">Lists</a></li>
               <li><a href="#lozenge">Lozenge</a></li>
-              <li><a href="#payment-card">Payment Card</a></li>
               <li><a href="#quote">Quote</a></li>
+            </ul>
+            <a href="#molecules">Molecules</a>
+            <ul>
+              <li><a href="#alert">Alert</a></li>
+              <li><a href="#card">Card</a></li>
+              <li><a href="#countdown">Countdown</a></li>
+              <li><a href="#payment-card">Payment Card</a></li>
               <li><a href="#rating">Rating</a></li>
-              <li><a href="#tile">Tile</a></li>
               <li><a href="#weather">Weather</a></li>
+            </ul>
+            <a href="#organisms">Organisms</a>
+            <ul>
             </ul>
           </li>
         </ul>
