+++ src/components/card/code/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./views/cardView.jsx');
