+++ src/components/card/code/less/_card.less
@@ -2,7 +2,7 @@
     #Tile
 \*------------------------------------*/
 
-.component-tile {
+.component-card {
   border-radius: @tiles-border-radius;
   border: @tiles-border-size @tiles-border-style @tiles-border-color;
   .caption {
