+++ src/components/card/code/templates/cardTemplate.jsx
@@ -3,7 +3,7 @@ var ImageComponent = require('../../../image');
 
 module.exports = function(props) {
   return (
-    <div className="component-tile">
+    <div className="component-card">
       <ImageComponent {...props.image} />
       <div className="caption">
         <h4>{props.title}</h4>
