--- src/components/justified-container/code/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./views/JustifiedContainerComponentView.jsx');
