--- src/components/justified-container/code/less/_justified-container.less
@@ -1,18 +1,0 @@
-/*------------------------------------*\
-    # Justified Container
-\*------------------------------------*/
-
-.component-justified-container {
-  display: table;
-  table-layout: fixed;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-
-  * {
-    flex:1;
-    display: table-cell;
-    width:100%;
-  }
-}
