--- src/components/justified-container/code/views/JustifiedContainerComponentView.jsx
@@ -1,8 +1,0 @@
-var React = require('react');
-
-module.exports = React.createClass({
-
-  render: function(){
-    return require('../templates/JustifiedContainerComponentTemplate.jsx').call(this);
-  }
-});
