--- src/components/tile/code/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./views/tileView.jsx');
