--- src/components/tile/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./code/index');
