+++ src/ui-toolkit.js
@@ -4,19 +4,18 @@ var UIToolkit = {};
 UIToolkit.Alert = require('./components/alert');
 UIToolkit.Anchor = require('./components/anchor');
 UIToolkit.Button = require('./components/button');
+UIToolkit.Card = require('./components/card');
 UIToolkit.Countdown = require('./components/countdown');
 UIToolkit.Flag = require('./components/flag');
 UIToolkit.IconList = require('./components/icon-list');
 UIToolkit.IconListItem = require('./components/icon-list-item');
 UIToolkit.Image = require('./components/image');
 UIToolkit.Input = require('./components/input');
-UIToolkit.JustifiedContainer = require('./components/justified-container');
 UIToolkit.Lozenge = require('./components/lozenge');
 UIToolkit.PaymentCard = require('./components/payment-card');
 UIToolkit.Quote = require('./components/quote');
 UIToolkit.Rating = require('./components/rating');
 UIToolkit.Reviews = require('./components/reviews');
-UIToolkit.Tile = require('./components/tile');
 UIToolkit.Weather = require('./components/weather');
 
 module.exports = UIToolkit;
